[SPARK-44530][CORE][CONNECT] Move SparkBuildInfo to common/util
### What changes were proposed in this pull request?
This PR moves `SparkBuildInfo` and the code that generates its properties to `common/util`.

### Why are the changes needed?
We need `SparkBuildInfo` in the connect scala client and we are removing connect's dependency on `core`.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Existing tests.

Closes #42133 from hvanhovell/SPARK-44530.

Authored-by: Herman van Hovell <herman@databricks.com>
